Arrange the code into appropriate files.
