+++ test/CMakeLists.txt
@@ -89,7 +89,7 @@ set(stdexec_test_sources
     exec/sequence/test_iterate.cpp
     exec/sequence/test_transform_each.cpp
     $<$<BOOL:${STDEXEC_ENABLE_TBB}>:tbbexec/test_tbb_thread_pool.cpp>
-    exec/test_system_scheduler.cpp
+    exec/test_system_context.cpp
     )
 
 add_executable(test.stdexec ${stdexec_test_sources})
